mjrt: Make library functions return i64 instead of i32
Eveything else is calculated in 64-bit registers, and this
discrepencay led to errors when comparing, especially to -1.

+++ mjrt-impl/src/lib.rs
@@ -13,7 +13,7 @@ extern "C" {
     pub fn mj_main();
 }
 
-pub type MjInt = i32;
+pub type MjInt = i64;
 
 macro_rules! mjrt_runtimeexception {
     ($fn_name:ident, $description:expr) => {
@@ -68,7 +68,7 @@ pub extern "C" fn mjrt_system_in_read() -> MjInt {
     let mut byte: [u8; 1] = [0];
 
     match stdin().read_exact(&mut byte) {
-        Ok(()) => i32::from(byte[0]),
+        Ok(()) => i64::from(byte[0]),
         Err(_) => -1,
     }
 }
